+++ examples/inference/cpp/decoder_example.cc.cu
