+++ examples/inference/cpp/gpt_generation.cc.cu
