+++ examples/inference/cpp/transformer_generate_example.cc.cu
