+++ examples/training/deepspeed/ds_fairseq_wmt14en2de.sh
@@ -11,7 +11,7 @@ fi
 
 deepspeed ${THIS_DIR}/ds_fairseq.py \
     ./wmt14_en_de/ \
-    --user-dir ./examples/fairseq/fs_modules \
+    --user-dir  ${THIS_DIR}/../../../lightseq/training/fs_modules \
     --arch ls_transformer_wmt_en_de_big_t2t --share-decoder-input-output-embed \
     --optimizer ls_adam --adam-betas '(0.9, 0.98)' --clip-norm 0.0 \
     --lr 5e-4 --lr-scheduler inverse_sqrt --warmup-updates 4000 \
