+++ setup.py
@@ -119,8 +119,8 @@ setup_kwargs = dict(
     },
     entry_points={
         "console_scripts": [
-            "lightseq-train = lightseq.training.examples.fairseq"
-            ".lightseq_fairseq_train_cli:ls_cli_main",
+            "lightseq-train = lightseq.training."
+            "lightseq_fairseq_train_cli:ls_cli_main",
         ],
     },
 )
