+++ tests/test_ls_kernels.py
@@ -4,7 +4,7 @@ import torch
 from torch.nn import functional
 
 from lightseq.training.ops.pytorch.builder import KernelBuilder, AdamBuilder
-from lightseq.training.tests.util import TestDecorator, cast_fp32_tensor
+from tests.util import TestDecorator, cast_fp32_tensor
 
 cuda_module = KernelBuilder().load()
 adam_module = AdamBuilder().load()
